Added Level 2 & Enemy Snake
Imported the map and changed colour to show that it's a different level.
Created the Enemy Snake and made it so it spawns from a food and has its own tail.
Up until now the enemy snake will not make the player loose.
Enemy always spawns from the same food