Obstacles
Added moving obstacles in level 3 and made them so the enemy snake would avoid the moving obstacles.
Made the enemy snake itself an obstacle as it was bugging while trying to avoid the obstacles.
Did not make it so that when the obstacles hit the player snake tail the game would end as it seemed impossible to get through